refactor(ivy): move all styling util code into `utils/styling_utils.ts`
