+++ packages/core/test/render3/styling_next/map_based_bindings_spec.ts
@@ -5,7 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {normalizeIntoStylingMap as createMap} from '../../../src/render3/styling/util';
+import {normalizeIntoStylingMap as createMap} from '../../../src/render3/util/styling_utils';
 
 describe('map-based bindings', () => {
   describe('StylingMapArray construction', () => {
