+++ packages/core/test/render3/styling_next/styling_context_spec.ts
@@ -7,9 +7,8 @@
  */
 import {registerBinding} from '@angular/core/src/render3/styling/bindings';
 import {attachStylingDebugObject} from '@angular/core/src/render3/styling/styling_debug';
-import {DEFAULT_GUARD_MASK_VALUE} from '@angular/core/src/render3/styling/util';
 
-import {allocTStylingContext} from '../../../src/render3/styling/util';
+import {DEFAULT_GUARD_MASK_VALUE, allocTStylingContext} from '../../../src/render3/util/styling_utils';
 
 describe('styling context', () => {
   it('should register a series of entries into the context', () => {
